test(spectating.spec): update casual spectate test to correctly assert new dialog flow
